+++ frontend/src/types/oauth.ts
@@ -1,0 +1,10 @@
+export type OAuthUserInfo = {
+  platformId: number;
+  email: string;
+  username: string;
+  profileImageUrl: string;
+};
+
+export type RequestOauthLoginBody = OAuthUserInfo & {
+  platformType: "KAKAO" | "NAVER" | "GOOGLE";
+};
